better hyperv testing

+++ data/test/vm/Test-HyperV.ps1
@@ -7,12 +7,13 @@ param(
     [string]$SerialOutputPath
 )
 
-# If we can't get the VMs, we're not running as admin
-$ErrorActionPreference = 'SilentlyContinue'
-$AllMachines = Get-VM
-$ErrorActionPreference = 'Continue'
+# Check if the user has permission to run Hyper-V VMs
+$LocalAdminGroup = [System.Security.Principal.SecurityIdentifier]::new("S-1-5-32-544")
+$HyperVAdminGroup = [System.Security.Principal.SecurityIdentifier]::new("S-1-5-32-578")
+$CurrentUser = [Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()
+$CanUserHyperV = $CurrentUser.IsInRole($LocalAdminGroup) -or $CurrentUser.IsInRole($HyperVAdminGroup)
 
-if ($AllMachines -eq $null) {
+if (!$CanUserHyperV) {
     Write-Output "This account does not have permission to run Hyper-V VMs. Please add your user to the Hyper-V Administrators group and try again."
     exit 0
 }
@@ -23,7 +24,7 @@ $ImagePath = $RootPath + '\bezos.iso'
 
 $VmName = 'Test-BezOS'
 $VhdPath = $RootPath + '\bezos.vhdx'
-$SerialPort = $RootPath + '\com1.txt'
+$SerialPipe = '\\.\pipe\SerialBezOS'
 
 $VM = @{
     Name = $VmName
@@ -41,9 +42,12 @@ New-Item -ItemType Directory -Path $RootPath -Force
 # Copy the kernel image to the hyperv directory
 Copy-Item -Path $KernelImage -Destination $ImagePath -Force
 
-Stop-VM -Name $VmName -Force -TurnOff
+$OldVm = Get-VM -Name $VmName -ErrorAction SilentlyContinue
+if ($OldVm) {
+    Stop-VM -Name $VmName -Force -TurnOff
 
-Remove-VM -Name $VmName -Force
+    Remove-VM -Name $VmName -Force
+}
 
 if (Test-Path $VhdPath) {
     Remove-Item $VhdPath -Force
@@ -53,7 +57,7 @@ $Vm = New-VM @VM
 
 $BootDrive = Add-VMDvdDrive -VMName $VmName -Path $ImagePath -Passthru
 
-# Set-VMComPort $VmName 1 -Path $SerialPort
+Set-VMComPort $VmName 1 -Path $SerialPipe
 
 Set-VMFirmware -VMName $VmName -FirstBootDevice $BootDrive -EnableSecureBoot Off
